refactor tests to check specific matchType upgrades
