+++ services/server/test/helpers/helpers.ts
@@ -1,3 +1,4 @@
+import config from "config";
 import {
   ContractFactory,
   Wallet,
@@ -12,7 +13,7 @@ import { assertVerificationSession, assertVerification } from "./assertions";
 import chai from "chai";
 import chaiHttp from "chai-http";
 import path from "path";
-import { promises as fs } from "fs";
+import { promises as fs, readFileSync } from "fs";
 import { ServerFixture } from "./ServerFixture";
 import type { Done } from "mocha";
 import { LocalChainFixture } from "./LocalChainFixture";
@@ -247,4 +248,125 @@ export async function resetDatabase(sourcifyDatabase: Pool) {
   await sourcifyDatabase.query("DELETE FROM compiled_contracts");
   await sourcifyDatabase.query("DELETE FROM contracts");
   await sourcifyDatabase.query("DELETE FROM code");
+}
+
+export async function testPartialUpgrade(
+  serverFixture: ServerFixture,
+  chainFixture: LocalChainFixture,
+  matchType: "creation" | "runtime",
+) {
+  const partialMetadata = (
+    await import("../testcontracts/Storage/metadataModified.json")
+  ).default;
+  const partialMetadataBuffer = Buffer.from(JSON.stringify(partialMetadata));
+
+  const partialSourcePath = path.join(
+    __dirname,
+    "..",
+    "testcontracts",
+    "Storage",
+    "StorageModified.sol",
+  );
+  const partialSourceBuffer = readFileSync(partialSourcePath);
+
+  let res = await chai
+    .request(serverFixture.server.app)
+    .post("/")
+    .field("address", chainFixture.defaultContractAddress)
+    .field("chain", chainFixture.chainId)
+    .field("creatorTxHash", chainFixture.defaultContractCreatorTx)
+    .attach("files", partialMetadataBuffer, "metadata.json")
+    .attach("files", partialSourceBuffer);
+  await assertVerification(
+    serverFixture.sourcifyDatabase,
+    null,
+    res,
+    null,
+    chainFixture.defaultContractAddress,
+    chainFixture.chainId,
+    "partial",
+  );
+
+  const contractMatchesWithPartialMetadata =
+    await serverFixture.sourcifyDatabase.query(
+      "SELECT runtime_match, creation_match FROM sourcify_matches;",
+    );
+
+  chai
+    .expect(contractMatchesWithPartialMetadata.rows[0].runtime_match)
+    .to.equal("partial");
+  chai
+    .expect(contractMatchesWithPartialMetadata.rows[0].creation_match)
+    .to.equal("partial");
+
+  const contractDeploymentWithoutCreatorTransactionHash =
+    await serverFixture.sourcifyDatabase.query(
+      "SELECT encode(transaction_hash, 'hex') as transaction_hash, block_number, transaction_index, contract_id FROM contract_deployments",
+    );
+  const contractIdWithoutCreatorTransactionHash =
+    contractDeploymentWithoutCreatorTransactionHash?.rows[0].contract_id;
+
+  // Force perfect ${matchType}Match by setting sourcify_match.${matchType}Match = "perfect" and moving contract to full_match
+  await serverFixture.sourcifyDatabase.query(
+    `UPDATE sourcify_matches SET ${matchType}_match='perfect' WHERE 1=1`,
+  );
+  const existingPath = path.join(
+    config.get("repositoryV1.path"),
+    "contracts",
+    "partial_match",
+    chainFixture.chainId,
+    chainFixture.defaultContractAddress,
+  );
+  const newPath = path.join(
+    config.get("repositoryV1.path"),
+    "contracts",
+    "full_match",
+    chainFixture.chainId,
+    chainFixture.defaultContractAddress,
+  );
+  await fs.mkdir(path.dirname(newPath), { recursive: true });
+  await fs.rename(existingPath, newPath);
+
+  // verify again with original metadata file
+  res = await chai
+    .request(serverFixture.server.app)
+    .post("/")
+    .field("address", chainFixture.defaultContractAddress)
+    .field("chain", chainFixture.chainId)
+    .field("creatorTxHash", chainFixture.defaultContractCreatorTx)
+    .attach("files", chainFixture.defaultContractMetadata, "metadata.json")
+    .attach("files", chainFixture.defaultContractSource);
+  await assertVerification(
+    serverFixture.sourcifyDatabase,
+    null,
+    res,
+    null,
+    chainFixture.defaultContractAddress,
+    chainFixture.chainId,
+  );
+
+  const contractMatchesWithPerfectMetadata =
+    await serverFixture.sourcifyDatabase.query(
+      "SELECT runtime_match, creation_match FROM sourcify_matches;",
+    );
+
+  chai
+    .expect(contractMatchesWithPerfectMetadata.rows[0].runtime_match)
+    .to.equal("perfect");
+  chai
+    .expect(contractMatchesWithPerfectMetadata.rows[0].creation_match)
+    .to.equal("perfect");
+
+  const contractDeploymentWithCreatorTransactionHash =
+    await serverFixture.sourcifyDatabase.query(
+      "SELECT encode(transaction_hash, 'hex') as transaction_hash, block_number, transaction_index, contract_id FROM contract_deployments",
+    );
+
+  const contractIdWithCreatorTransactionHash =
+    contractDeploymentWithCreatorTransactionHash?.rows[0].contract_id;
+
+  // There should not be a new contract_id
+  chai
+    .expect(contractIdWithCreatorTransactionHash)
+    .to.equal(contractIdWithoutCreatorTransactionHash);
 }
